Structure Visitor, Visitable, VisitableAdapter, Repository established
